+++ prueba/biblioPropia.h
@@ -41,9 +41,29 @@ int ingresarString(char mensaje[], char info[]);
  */
 int pasarMayusculaPrimerCaracter(char string[]);
 
+/** \brief valida  que un numero entero este dentro de un rango entre minimo y maximo
+ *
+ * \param numero int el numero a validar
+ * \param min int el minimo
+ * \param max int el maximo
+ * \return int retorna 0 si el numero esta fuera de rango o 1 si esta ok
+ *
+ */
 int validarInt(int numero,int min,int max);
 
-float validarFloat(float numero,int min,int max);
+
+/** \brief valida que un numero decimal este dentro de un rango entre minimo y maximo
+ *
+ * \param numero float el numero a validar
+ * \param min int el minimo de rango
+ * \param max int el maximo de rango
+ * \return int 0 si esta fuera de rango o 1 si esta ok
+ *
+ */
+int validarFloat(float numero,int min,int max);
+
+
+
 
 int orden();
